+++ src/main/java/com/tj/kvasir/web/rest/CategorySubjectResource.java
@@ -1,14 +1,20 @@
 package com.tj.kvasir.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
+import com.tj.kvasir.domain.CategoryNode;
 import com.tj.kvasir.domain.CategorySubject;
 
+import com.tj.kvasir.domain.enumeration.CategoryType;
+import com.tj.kvasir.repository.CategoryNodeRepository;
 import com.tj.kvasir.repository.CategorySubjectRepository;
 import com.tj.kvasir.repository.search.CategorySubjectSearchRepository;
 import com.tj.kvasir.web.rest.util.HeaderUtil;
 import io.github.jhipster.web.util.ResponseUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.data.domain.Example;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -38,9 +44,12 @@ public class CategorySubjectResource {
 
     private final CategorySubjectSearchRepository categorySubjectSearchRepository;
 
-    public CategorySubjectResource(CategorySubjectRepository categorySubjectRepository, CategorySubjectSearchRepository categorySubjectSearchRepository) {
+    private final CategoryNodeRepository categoryNodeRepository;
+
+    public CategorySubjectResource(CategorySubjectRepository categorySubjectRepository, CategorySubjectSearchRepository categorySubjectSearchRepository, CategoryNodeRepository categoryNodeRepository) {
         this.categorySubjectRepository = categorySubjectRepository;
         this.categorySubjectSearchRepository = categorySubjectSearchRepository;
+        this.categoryNodeRepository = categoryNodeRepository;
     }
 
     /**
@@ -123,9 +132,18 @@ public class CategorySubjectResource {
     @Timed
     public ResponseEntity<Void> deleteCategorySubject(@PathVariable Long id) {
         log.debug("REST request to delete CategorySubject : {}", id);
-        categorySubjectRepository.delete(id);
-        categorySubjectSearchRepository.delete(id);
-        return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        CategoryNode node = new CategoryNode();
+        node.setType(CategoryType.SUBJECT);
+        node.setTypeId(id);
+        List<CategoryNode> nodes = categoryNodeRepository.findAll(Example.of(node));
+        if (nodes.isEmpty()) {
+            categorySubjectRepository.delete(id);
+            categorySubjectSearchRepository.delete(id);
+            return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        } else {
+            HttpHeaders alert = HeaderUtil.createFailureAlert(ENTITY_NAME, "categoryinuse", "Can not delete category in use");
+            return ResponseEntity.status(HttpStatus.CONFLICT).headers(alert).build();
+        }
     }
 
     /**
